feat(part4/blogs): exercise 4.22 completed

+++ part4/blogs/controllers/blogs.js
@@ -1,5 +1,8 @@
 const blogsRouter = require("express").Router();
 const Blog = require("../models/blog");
+const User = require("../models/user");
+
+const jwt = require("jsonwebtoken");
 
 blogsRouter.get("/", async (request, response) => {
   const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 });
@@ -11,19 +14,44 @@ blogsRouter.post("/", async (request, response) => {
     return response.status(400);
   }
 
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token invalid" });
+  }
+
+  const user = request.user;
+
   const payload = {
     ...request.body,
     likes: request.body.likes ? request.body.likes : 0,
+    user: user._id,
   };
 
   const blog = new Blog(payload);
   const result = await blog.save();
+
+  user.blogs = user.blogs.concat(result._id);
+  await user.save();
+
   response.status(201).json(result);
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
   if (!request.params.id) {
     return response.status(400);
+  }
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token invalid" });
+  }
+
+  const user = await request.user;
+  const blog = await Blog.findById(request.params.id);
+
+  if (blog.user.toString() !== user._id.toString()) {
+    return response.status(401).json({ error: "Operation denied" });
   }
 
   await Blog.findByIdAndRemove(request.params.id);
